[FEATURE] Add firebase initial configuration. - [x] Remove some files/folders from the template.
