+++ package.json
@@ -9,12 +9,14 @@
     "@material-ui/icons": "4.2.1",
     "chartist": "0.10.1",
     "classnames": "2.2.6",
+    "firebase": "^7.14.2",
     "history": "4.9.0",
     "perfect-scrollbar": "1.4.0",
     "prop-types": "15.7.2",
     "react": "16.9.0",
     "react-chartist": "0.13.3",
     "react-dom": "16.9.0",
+    "react-firebase-hooks": "^2.1.1",
     "react-google-maps": "9.4.5",
     "react-router-dom": "5.0.1",
     "react-scripts": "^3.4.1",
