+++ src/index.js
@@ -1,24 +1,9 @@
-/*!
-
-=========================================================
-* Material Dashboard React - v1.8.0
-=========================================================
-
-* Product Page: https://www.creative-tim.com/product/material-dashboard-react
-* Copyright 2019 Creative Tim (https://www.creative-tim.com)
-* Licensed under MIT (https://github.com/creativetimofficial/material-dashboard-react/blob/master/LICENSE.md)
-
-* Coded by Creative Tim
-
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-*/
 import React from "react";
 import ReactDOM from "react-dom";
 import { createBrowserHistory } from "history";
 import { Router, Route, Switch, Redirect } from "react-router-dom";
+import firebase from "firebase";
+import { firebaseConfig } from "./config/firebase";
 
 // core components
 import Admin from "layouts/Admin.js";
@@ -27,6 +12,7 @@ import RTL from "layouts/RTL.js";
 import "assets/css/material-dashboard-react.css?v=1.8.0";
 
 const hist = createBrowserHistory();
+firebase.initializeApp(firebaseConfig);
 
 ReactDOM.render(
   <Router history={hist}>
