+++ src/views/ProjectList/ProjectList.js
@@ -1,0 +1,80 @@
+import React from "react";
+// @material-ui/core components
+import { makeStyles } from "@material-ui/core/styles";
+// core components
+import GridItem from "components/Grid/GridItem.js";
+import GridContainer from "components/Grid/GridContainer.js";
+import Table from "components/Table/Table.js";
+import Card from "components/Card/Card.js";
+import CardHeader from "components/Card/CardHeader.js";
+import CardBody from "components/Card/CardBody.js";
+import { useCollection } from 'react-firebase-hooks/firestore';
+import firebase from "firebase";
+
+const styles = {
+  cardCategoryWhite: {
+    "&,& a,& a:hover,& a:focus": {
+      color: "rgba(255,255,255,.62)",
+      margin: "0",
+      fontSize: "14px",
+      marginTop: "0",
+      marginBottom: "0"
+    },
+    "& a,& a:hover,& a:focus": {
+      color: "#FFFFFF"
+    }
+  },
+  cardTitleWhite: {
+    color: "#FFFFFF",
+    marginTop: "0px",
+    minHeight: "auto",
+    fontWeight: "300",
+    fontFamily: "'Roboto', 'Helvetica', 'Arial', sans-serif",
+    marginBottom: "3px",
+    textDecoration: "none",
+    "& small": {
+      color: "#777",
+      fontSize: "65%",
+      fontWeight: "400",
+      lineHeight: "1"
+    }
+  }
+};
+
+const useStyles = makeStyles(styles);
+
+export default function ProjectList() {
+  const [data, loading, error] = useCollection(
+    firebase.firestore().collection('Project'),
+    {
+      snapshotListenOptions: { includeMetadataChanges: true },
+    }
+  );
+  const classes = useStyles();
+  return (
+    <GridContainer>
+      <GridItem xs={12} sm={12} md={12}>
+        <Card>
+          <CardHeader color="primary">
+            <h4 className={classes.cardTitleWhite}>Projects</h4>
+            <p className={classes.cardCategoryWhite}>
+               List of projects
+            </p>
+          </CardHeader>
+          <CardBody>
+            {data && data.docs.map(doc => (
+              <React.Fragment key={doc.id}>
+                {JSON.stringify(doc.data())},{' '}
+              </React.Fragment>
+            ))}
+            <Table
+              tableHeaderColor="primary"
+              tableHead={["Name", "Description", "Status", "Comments", "Votes"]}
+              tableData={loading ? [] : []}
+            />
+          </CardBody>
+        </Card>
+      </GridItem>
+    </GridContainer>
+  );
+}
