--- votatron-dashboard-react-master/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": ["es2015"],
-  "plugins": [
-    "transform-class-properties",
-    "transform-react-jsx",
-    "transform-object-rest-spread",
-    ["module-resolver", {
-      "root": ["./src"]
-      }],
-      ["import-rename", {"^(.*)\\.jsx$": "$1"}]
-    ]
-  }
