--- votatron-dashboard-react-master/.eslintignore
@@ -1,3 +1,0 @@
-!.eslintrc.js
-documentation/
-build/
